+++ CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.54.1](https://github.com/rokucommunity/brighterscript/compare/v0.54.0...0.54.1) - 2022-07-22
+### Changed
+ - Fix bug in languageserver hover provider ([#649](https://github.com/rokucommunity/brighterscript/pull/649))
+
+
+
+## [0.54.0](https://github.com/rokucommunity/brighterscript/compare/v0.53.1...0.54.0) - 2022-07-22
+### Added
+ - hover for CONST references. ([#648](https://github.com/rokucommunity/brighterscript/pull/648))
+ - plugins can contribute completions ([#647](https://github.com/rokucommunity/brighterscript/pull/647))
+ - plugins can contribute hovers ([#393](https://github.com/rokucommunity/brighterscript/pull/393))
+### Changed
+ - Export some vscode interfaces ([#644](https://github.com/rokucommunity/brighterscript/pull/644))
+ - Better plugin docs ([#643](https://github.com/rokucommunity/brighterscript/pull/643))
+
+
+
 ## [0.53.1](https://github.com/rokucommunity/brighterscript/compare/v0.53.0...0.53.1) - 2022-07-15
 ### Fixed
  - bug with codeactions at the edges of tokens. ([#642](https://github.com/rokucommunity/brighterscript/pull/642))